preprocessor (1).py
"""
Advanced Image Preprocessing with Augmentation Simulation
"""

import numpy as np
from PIL import Image

class AdvancedImagePreprocessor:
    def __init__(self, target_size=(224, 224)):
        self.target_size = target_size
        self.augmentation_techniques = self._get_augmentation_techniques()
    
    def _get_augmentation_techniques(self):
        return {
            'geometric': ['random_flip', 'rotation', 'zoom', 'shear', 'translation'],
            'photometric': ['brightness', 'contrast', 'saturation', 'hue', 'color_jitter'],
            'advanced': ['mixup', 'cutmix', 'random_erasing', 'gaussian_noise']
        }
    
    def advanced_preprocess(self, image, augment=False):
        """Advanced preprocessing pipeline"""
        # Resize with anti-aliasing
        img = image.resize(self.target_size, Image.Resampling.LANCZOS)
        
        # Convert to array
        img_array = np.array(img, dtype=np.float32) / 255.0
        
        # Advanced normalization (ImageNet stats)
        mean = [0.485, 0.456, 0.406]
        std = [0.229, 0.224, 0.225]
        
        if len(img_array.shape) == 3:
            for i in range(3):
                img_array[:, :, i] = (img_array[:, :, i] - mean[i]) / std[i]
        
        # Simulate augmentation
        if augment:
            img_array = self._simulate_augmentation(img_array)
        
        # Add batch dimension
        img_array = np.expand_dims(img_array, axis=0)
        
        return img_array
    
    def _simulate_augmentation(self, img_array):
        """Simulate data augmentation effects"""
        # Simulate random flip
        if np.random.random() > 0.5:
            img_array = np.fliplr(img_array)
        
        # Simulate brightness adjustment
        brightness_factor = np.random.uniform(0.8, 1.2)
        img_array = np.clip(img_array * brightness_factor, 0, 1)
        
        return img_array
    
    def extract_advanced_features(self, image):
        """Extract advanced image features"""
        img_array = np.array(image)
        
        features = {
            'color_moments': self._calculate_color_moments(img_array),
            'texture_features': self._calculate_texture_features(img_array),
            'edge_statistics': self._calculate_edge_stats(img_array),
            'shape_descriptors': self._calculate_shape_descriptors(img_array)
        }
        
        return features
    
    def _calculate_color_moments(self, img_array):
        """Calculate color moments (mean, std, skewness)"""
        if len(img_array.shape) == 3:
            moments = []
            for channel in range(3):
                channel_data = img_array[:, :, channel].flatten()
                mean = np.mean(channel_data)
                std = np.std(channel_data)
                skewness = np.mean((channel_data - mean) ** 3) / (std ** 3)
                moments.extend([mean, std, skewness])
            return moments
        return []
    
    def _calculate_texture_features(self, img_array):
        """Calculate texture features"""
        if len(img_array.shape) == 3:
            gray = np.mean(img_array, axis=2)
        else:
            gray = img_array
        
        return {
            'contrast': np.std(gray),
            'entropy': -np.sum(gray * np.log2(gray + 1e-8)),
            'smoothness': 1 - 1/(1 + np.var(gray))
        }
    
    def _calculate_edge_stats(self, img_array):
        """Calculate edge statistics"""
        if len(img_array.shape) == 3:
            gray = np.mean(img_array, axis=2)
        else:
            gray = img_array
        
        # Simple edge detection simulation
        gradient_x = np.diff(gray, axis=1)
        gradient_y = np.diff(gray, axis=0)
        
        return {
            'edge_density': np.mean(np.abs(gradient_x)) + np.mean(np.abs(gradient_y)),
            'edge_variance': np.var(gradient_x) + np.var(gradient_y)
        }
    
    def _calculate_shape_descriptors(self, img_array):
        """Calculate shape descriptors"""
        if len(img_array.shape) == 3:
            gray = np.mean(img_array, axis=2)
        else:
            gray = img_array
        
        return {
            'aspect_ratio': gray.shape[1] / gray.shape[0],
            'area_ratio': np.mean(gray > np.mean(gray))
        }

def demonstrate_advanced_preprocessing():
    """Demonstrate advanced preprocessing capabilities"""
    preprocessor = AdvancedImagePreprocessor()
    
    print("🔧 ADVANCED PREPROCESSING PIPELINE")
    print("=" * 50)
    print("Target Size:", preprocessor.target_size)
    
    print("\n🔄 AUGMENTATION TECHNIQUES:")
    for category, techniques in preprocessor.augmentation_techniques.items():
        print(f"  {category.title()}: {', '.join(techniques)}")
    
    print("\n📊 FEATURE EXTRACTION:")
    features = preprocessor.extract_advanced_features(Image.new('RGB', (100, 100)))
    for feature_type, feature_data in features.items():
        print(f"  {feature_type}: {len(feature_data) if isinstance(feature_data, list) else len(feature_data.keys())} metrics")

if __name__ == "__main__":
    demonstrate_advanced_preprocessing()